installing multer
-adding multer the next middlware that im using to save the image inside of my images folder
-trying to fix the auth part where the connected user doesnt have the rights to add new book same as the one not authenticated

+++ backend/controllers/books.js
@@ -1,20 +1,24 @@
 const Book = require('../models/book');
 
 exports.createBook = (req, res, next) => {
+  //we want to get the stringified book and parse to turn into json obj
+  req.body.book = JSON.parse(req.body.book)
+  //build URL
+  const url = req.protocol + '://' + req.get('host');
   const book = new Book({
-    userId: req.body.userId,
-    title: req.body.title,
-    author: req.body.author,
-    imageUrl: req.body.imageUrl,
-    year: req.body.year,
-    genre: req.body.genre,
+    userId: req.auth.userId,
+    title: req.body.book.title,
+    author: req.body.book.author,
+    imageUrl: url + '/images/' + req.file.filename,
+    year: Number(req.body.book.year),
+    genre: req.body.book.genre,
     ratings: [
       {
-        userId: req.body.userId,
-        grade: req.body.grade
+        userId: req.auth.userId,
+        grade: Number(req.body.book.grade)
       }
     ],
-    averageRating: req.body.averageRating
+    averageRating: Number(req.body.book.averageRating)
   });
   book.save().then(
     () => {
@@ -48,21 +52,49 @@ exports.getOneBook = (req, res, next) => {
 };
 
 exports.modifyBook = (req, res, next) => {
-  const book = new Book({
-    userId: req.auth.userId,
-    title: req.body.title,
-    author: req.body.author,
-    imageUrl: req.body.imageUrl,
-    year: req.body.year,
-    genre: req.body.genre,
-    ratings: [
-      {
-        userId: req.auth.userId,
-        grade: req.body.grade
-      }
-    ],
-    averageRating: req.body.averageRating
-  });
+  //creating new book will create new id so we will pass the id 
+  let book = new Book({ _id: req.params.id });
+  //check if there is a file with the request
+  if (req.file) {
+    //we want to get the stringified book and parse to turn into json obj
+    req.body.book = JSON.parse(req.body.book)
+    //build URL
+    const url = req.protocol + '://' + req.get('host');
+    book = {
+      //pass the same id
+      _id: req.params.id,
+      userId: req.auth.userId,
+      title: req.body.book.title,
+      author: req.body.book.author,
+      imageUrl: url + '/images/' + req.file.filename,
+      year: Number(req.body.book.year),
+      genre: req.body.book.genre,
+      ratings: [
+        {
+          userId: req.auth.userId,
+          grade: Number(req.body.book.grade)
+        }
+      ],
+      averageRating: Number(req.body.book.averageRating)
+    };
+  } else {
+    book = {
+      _id: req.params.id,
+      userId: req.auth.userId,
+      title: req.body.title,
+      author: req.body.author,
+      imageUrl: req.body.imageUrl,
+      year: Number(req.body.year),
+      genre: req.body.genre,
+      ratings: [
+        {
+          userId: req.auth.userId,
+          grade: Number(req.body.grade)
+        }
+      ],
+      averageRating: Number(req.body.averageRating)
+    };
+  }
   Book.updateOne({ _id: req.params.id }, book).then(
     () => {
       res.status(201).json({
@@ -79,19 +111,33 @@ exports.modifyBook = (req, res, next) => {
 }
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id }).then(
-    () => {
-      res.status(200).json({
-        message: 'Deleted!'
-      });
+  Book.findOne({ _id: req.params.id }).then(
+    (book) => {
+      if (!book) {
+        return res.status(404).json({
+          error: new Error('There is no book')
+        });
+      }
+      if (book.userId !== req.auth.userId) {
+        return res.status(400).json({
+          erorr: new Error('Unauthorized request!!!')
+        });
+      }
+      Book.deleteOne({ _id: req.params.id }).then(
+        () => {
+          res.status(200).json({
+            message: 'Deleted!'
+          });
+        }
+      ).catch(
+        (error) => {
+          res.status(400).json({
+            error: error
+          });
+        }
+      );
     }
-  ).catch(
-    (error) => {
-      res.status(400).json({
-        error: error
-      });
-    }
-  );
+  )
 };
